TIMOB-15267-Moved the apis to Ti.App.Android
