+++ apidoc/Titanium/Media/Android/Android.yml
@@ -52,18 +52,6 @@ methods:
             If `true`, the image is scaled to fit the width of the system wallpaper. If
             `false`, the image is not scaled.
         type: Boolean
-properties:
-  - name: appVersionCode
-    summary: |
-        The version number of the application. 
-    type: Number
-    since: 3.3.0
-    
-  - name: appVersionName
-    summary: |
-        The version name of the application. 
-    type: Number
-    since: 3.3.0
 
 ---
 name: MediaScannerResponse
